GH-23 Move methods to node for better interface
