+++ rosbag2/CMakeLists.txt
@@ -110,14 +110,6 @@ if(BUILD_TESTING)
     )
   endif()
 
-  ament_add_gmock(rosbag2_integration_test
-    test/rosbag2/rosbag2_integration_test.cpp
-    src/rosbag2/typesupport_helpers.cpp
-    WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
-  if(TARGET rosbag2_integration_test)
-    target_link_libraries(rosbag2_integration_test librosbag2)
-  endif()
-
   ament_add_gmock(rosbag2_rosbag_node_test
     test/rosbag2/rosbag2_rosbag_node_test.cpp
     test/rosbag2/test_memory_management.cpp
